The coordinates can be completely read by county and the max + min can be read in for the bounding box.
